Refactor interfaces
